+++ hello/src/main/java/com/hello/hello/StudentForm.java
@@ -8,7 +8,6 @@ import java.time.LocalDate;
 
 public class StudentForm {
     @Pattern(regexp = "^[0-9]{6}$")
-    @NotBlank
     private String studentId;
 
     @NotBlank
