+++ hello/src/main/java/com/hello/hello/StudentsController.java
@@ -1,9 +1,7 @@
 package com.hello.hello;
 
-import jakarta.validation.ConstraintViolationException;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -11,11 +9,9 @@ import org.springframework.web.util.UriComponentsBuilder;
 
 import java.net.URI;
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-@RestControllerAdvice
 @RestController
 @RequestMapping("/students")
 @Validated
@@ -45,13 +41,13 @@ public class StudentsController {
                 .filter(form -> form.getName().equals(name))
                 .findFirst()
                 .orElse(null);
-        // クエリパラメータがnullまたは空白の場合
+        // クエリパラメータがnullの場合
         if (studentList == null) {
             return new String("入力された値は不適当です");
         }
         // クエリパラメータで指定されたnameが存在する場合
         else {
-            return studentList.getName() + new String("さんIDを入力してください");
+            return studentList.getName() + new String("さんようこそ！");
         }
     }
 
@@ -78,40 +74,5 @@ public class StudentsController {
     public ResponseEntity<Map<String, String>> delete(@PathVariable("studentId") int id) {
         // 削除処理省略
         return ResponseEntity.ok(Map.of("message", "studentId successfully deleted"));
-    }
-
-    public static final class ErrorResponse {
-        private final HttpStatus status;
-        private final String message;
-        private final List<Map<String, String>> errors;
-
-        public ErrorResponse(HttpStatus status, String message, List<Map<String, String>> errors) {
-            this.status = status;
-            this.message = message;
-            this.errors = errors;
-        }
-
-        public HttpStatus getStatus() {
-            return status;
-        }
-
-        public String getMessage() {
-            return message;
-        }
-
-        public List<Map<String, String>> getErrors() {
-            return errors;
-        }
-    }
-
-    @ExceptionHandler(value = ConstraintViolationException.class)
-    public ResponseEntity<Object> handleConstraintViolationException(ConstraintViolationException ex) {
-        List<Map<String, String>> errors = new ArrayList<>();
-        for (var violation : ex.getConstraintViolations()) {
-            errors.add(Map.of("key", violation.getPropertyPath().toString(), "message", violation.getMessage()));
-        }
-        ErrorResponse errorResponse =
-                new ErrorResponse(HttpStatus.BAD_REQUEST, "validation error", errors);
-        return ResponseEntity.badRequest().body(errorResponse);
     }
 }
